refactor: separated animated view from swiper
