+++ src/app/Scenes/InfiniteDiscovery/Components/Swiper/useScreenWidthWithOffset.ts
@@ -1,0 +1,6 @@
+import { useScreenDimensions } from "@artsy/palette-mobile"
+
+export const useScreenWidthWithOffset = () => {
+  const { width } = useScreenDimensions()
+  return width + 100
+}
